+++ src/cli/chain/create/create.js
@@ -2,28 +2,32 @@ const chalk = require('chalk');
 const fs = require('fs');
 const toml = require('@iarna/toml');
 const { config } = require('core/env');
-const { setConfigToProfile } = require('core/forge-config');
+const { setConfigToChain } = require('core/forge-config');
 const { printInfo, printError } = require('core/util');
 const {
-  createNewProfile,
+  createNewChain,
   getOriginForgeReleaseFilePath,
-  getProfileReleaseFilePath,
+  getChainReleaseFilePath,
 } = require('core/forge-fs');
 const { askUserConfigs, writeConfigs } = require('../../node/config/lib');
 
-async function main({ args: [chainName = ''] }) {
+async function main({ args: [chainName = ''], opts: { defaults, allowMultiChain = false } }) {
+  if (allowMultiChain === false) {
+    printError('Forge CLI is configured to work with single chain only, abort!');
+    process.exit(0);
+  }
   try {
     let configs = toml.parse(
       fs.readFileSync(getOriginForgeReleaseFilePath(config.get('cli').currentVersion)).toString()
     );
-    configs = await askUserConfigs(configs, chainName, true);
+    configs = await askUserConfigs(configs, chainName, { interactive: !defaults, isCreate: true });
 
     const {
       app: { name },
     } = configs;
-    configs = await setConfigToProfile(configs, name);
-    createNewProfile(name);
-    await writeConfigs(getProfileReleaseFilePath(name), configs);
+    configs = await setConfigToChain(configs, name);
+    createNewChain(name);
+    await writeConfigs(getChainReleaseFilePath(name), configs);
     printInfo(`Run ${chalk.cyan(`forge start ${name}`)} to start the chain`);
   } catch (error) {
     printError('Create new chain failed:');