+++ src/cli/project/create/index.js
@@ -16,10 +16,8 @@ cli(
       wallet: false,
     },
     options: [
-      ['-y, --yes', 'Use default value for all options'],
       ['--target <target>', 'Target directory, default is current directory'],
       ['--starterDir <starterDir>', 'Local boilerplate directory'],
-      ['--registry <registry>', 'Only for npm package'],
     ],
     handlers: {
       '--help': () => {