+++ src/core/cli.js
@@ -1,10 +1,13 @@
 /* eslint no-console:"off" */
 const path = require('path');
 const last = require('lodash/last');
+const rcfile = require('rcfile');
+const registryUrl = require('registry-url');
 
 const { setupEnv } = require('./env');
 
 const allCommands = [];
+const globalConfig = rcfile('forge');
 
 /**
  * create a cli
@@ -72,11 +75,19 @@ function initCli(program) {
 
       command.action(async (...params) => {
         const globalArgs = last(program.args).parent;
+        const globalOpts = Object.assign(
+          { allowMultiChain: true },
+          globalConfig,
+          globalArgs.opts()
+        );
+        if (globalOpts.registry === undefined) {
+          globalOpts.registry = registryUrl();
+        }
 
-        await setupEnv(globalArgs.args, x.requirements, globalArgs.opts());
+        await setupEnv(globalArgs.args, x.requirements, globalOpts);
         await x.handler({
           args: params.filter(p => typeof p === 'string'),
-          opts: Object.assign(globalArgs.opts(), command.opts()),
+          opts: Object.assign(globalOpts, command.opts()),
         });
       });
     });