+++ src/core/forge-config.js
@@ -4,7 +4,6 @@ const semver = require('semver');
 const path = require('path');
 const { get, set } = require('lodash');
 const TOML = require('@iarna/toml');
-const yaml = require('yaml');
 
 const {
   DEFAULT_CHAIN_NAME,
@@ -19,9 +18,11 @@ const {
   getCurrentReleaseFilePath,
   getOriginForgeReleaseFilePath,
   getForgeVersionFromYaml,
-  getProfileWorkshopDirectory,
-  getProfileDirectory,
-  getProfileReleaseFilePath,
+  getChainConfigPath,
+  getChainWorkshopDirectory,
+  getChainDirectory,
+  getChainReleaseFilePath,
+  updateChainConfig,
   requiredDirs,
 } = require('./forge-fs');
 const {
@@ -52,9 +53,9 @@ async function getUsedPortsByForge() {
   const maxPortPerField = {
     forgeWebPort: 0,
     forgeGrpcPort: 0,
-    tendminRpcPort: 0,
-    tendmintGrpcPort: 0,
-    tendmintP2pPort: 0,
+    tendermintRpcPort: 0,
+    tendermintGrpcPort: 0,
+    tendermintP2pPort: 0,
     workshopPort: 0,
   };
 
@@ -74,9 +75,9 @@ async function getUsedPortsByForge() {
     const forgeWebPort = Number(get(cfg, 'forge.web.port'));
     const workshopPort = Number(get(cfg, 'workshop.port'));
     const forgeGrpcPort = getPortFromUri(get(cfg, 'forge.sock_grpc'));
-    const tendminRpcPort = getPortFromUri(get(cfg, 'tendermint.sock_rpc'));
-    const tendmintGrpcPort = getPortFromUri(get(cfg, 'tendermint.sock_grpc'));
-    const tendmintP2pPort = getPortFromUri(get(cfg, 'tendermint.sock_p2p'));
+    const tendermintRpcPort = getPortFromUri(get(cfg, 'tendermint.sock_rpc'));
+    const tendermintGrpcPort = getPortFromUri(get(cfg, 'tendermint.sock_grpc'));
+    const tendermintP2pPort = getPortFromUri(get(cfg, 'tendermint.sock_p2p'));
 
     if (forgeWebPort > maxPortPerField.forgeWebPort) {
       maxPortPerField.forgeWebPort = forgeWebPort;
@@ -90,15 +91,15 @@ async function getUsedPortsByForge() {
       maxPortPerField.forgeGrpcPort = forgeGrpcPort;
     }
 
-    if (tendminRpcPort > maxPortPerField.tendminRpcPort) {
-      maxPortPerField.tendminRpcPort = tendminRpcPort;
+    if (tendermintRpcPort > maxPortPerField.tendermintRpcPort) {
+      maxPortPerField.tendermintRpcPort = tendermintRpcPort;
     }
-    if (tendmintGrpcPort > maxPortPerField.tendmintGrpcPort) {
-      maxPortPerField.tendmintGrpcPort = tendmintGrpcPort;
+    if (tendermintGrpcPort > maxPortPerField.tendermintGrpcPort) {
+      maxPortPerField.tendermintGrpcPort = tendermintGrpcPort;
     }
 
-    if (tendmintP2pPort > maxPortPerField.tendmintP2pPort) {
-      maxPortPerField.tendmintP2pPort = tendmintP2pPort;
+    if (tendermintP2pPort > maxPortPerField.tendermintP2pPort) {
+      maxPortPerField.tendermintP2pPort = tendermintP2pPort;
     }
   });
 
@@ -109,9 +110,9 @@ async function getAvailablePort() {
   const {
     forgeWebPort,
     forgeGrpcPort,
-    tendminRpcPort,
-    tendmintGrpcPort,
-    tendmintP2pPort,
+    tendermintRpcPort,
+    tendermintGrpcPort,
+    tendermintP2pPort,
     workshopPort,
   } = await getUsedPortsByForge();
 
@@ -119,14 +120,14 @@ async function getAvailablePort() {
     forgeWebPort: forgeWebPort
       ? forgeWebPort + 1
       : await getPort({ port: getPort.makeRange(8211, 8300) }),
-    tendminRpcPort: tendminRpcPort
-      ? tendminRpcPort + 1
+    tendermintRpcPort: tendermintRpcPort
+      ? tendermintRpcPort + 1
       : await getPort({ port: getPort.makeRange(32001, 34000) }),
-    tendmintGrpcPort: tendmintGrpcPort
-      ? tendmintGrpcPort + 1
+    tendermintGrpcPort: tendermintGrpcPort
+      ? tendermintGrpcPort + 1
       : await getPort({ port: getPort.makeRange(36001, 37000) }),
-    tendmintP2pPort: tendmintP2pPort
-      ? tendmintP2pPort + 1
+    tendermintP2pPort: tendermintP2pPort
+      ? tendermintP2pPort + 1
       : await getPort({ port: getPort.makeRange(37001, 38000) }),
     forgeGrpcPort: forgeGrpcPort
       ? forgeGrpcPort + 1
@@ -142,15 +143,22 @@ async function getAvailablePort() {
 function setConfig(
   configs,
   chainName,
-  { forgeWebPort, forgeGrpcPort, tendminRpcPort, tendmintGrpcPort, tendmintP2pPort, workshopPort }
+  {
+    forgeWebPort,
+    forgeGrpcPort,
+    tendermintRpcPort,
+    tendermintGrpcPort,
+    tendermintP2pPort,
+    workshopPort,
+  }
 ) {
   let content = JSON.parse(JSON.stringify(configs));
 
   set(content, 'forge.web.port', forgeWebPort);
   set(content, 'forge.sock_grpc', `tcp://127.0.0.1:${forgeGrpcPort}`);
-  set(content, 'tendermint.sock_rpc', `tcp://127.0.0.1:${tendminRpcPort}`);
-  set(content, 'tendermint.sock_grpc', `tcp://127.0.0.1:${tendmintGrpcPort}`);
-  set(content, 'tendermint.sock_p2p', `tcp://0.0.0.0:${tendmintP2pPort}`);
+  set(content, 'tendermint.sock_rpc', `tcp://127.0.0.1:${tendermintRpcPort}`);
+  set(content, 'tendermint.sock_grpc', `tcp://127.0.0.1:${tendermintGrpcPort}`);
+  set(content, 'tendermint.sock_p2p', `tcp://0.0.0.0:${tendermintP2pPort}`);
   set(content, 'workshop.port', workshopPort);
   set(content, 'workshop.local_forge', `tcp://127.0.0.1:${forgeGrpcPort}`);
 
@@ -164,34 +172,34 @@ function setFilePathOfConfig(configs, chainName) {
   const releaseDirectory = getDataDirectory(chainName);
 
   set(content, 'forge.path', path.join(releaseDirectory, 'core'));
-  set(content, 'tendermint.keypath', path.join(getProfileDirectory(chainName), 'keys'));
+  set(content, 'tendermint.keypath', path.join(getChainDirectory(chainName), 'keys'));
   set(content, 'tendermint.path', path.join(releaseDirectory, 'tendermint'));
 
   set(content, 'ipfs.path', path.join(releaseDirectory, 'ipfs'));
   set(content, 'cache.path', path.join(releaseDirectory, 'cache', 'mnesia_data_dir'));
 
-  set(content, 'workshop.path', getProfileWorkshopDirectory(chainName));
+  set(content, 'workshop.path', getChainWorkshopDirectory(chainName));
   set(content, 'workshop.db', 'sqlite://workshop.sqlite3');
 
   return content;
 }
 
-async function setConfigToProfile(configs, chainName) {
+async function setConfigToChain(configs, chainName) {
   const {
     forgeWebPort,
     forgeGrpcPort,
-    tendminRpcPort,
-    tendmintGrpcPort,
-    tendmintP2pPort,
+    tendermintRpcPort,
+    tendermintGrpcPort,
+    tendermintP2pPort,
     workshopPort,
   } = await getAvailablePort();
 
   return setConfig(configs, chainName, {
     forgeWebPort,
     forgeGrpcPort,
-    tendminRpcPort,
-    tendmintGrpcPort,
-    tendmintP2pPort,
+    tendermintRpcPort,
+    tendermintGrpcPort,
+    tendermintP2pPort,
     workshopPort,
   });
 }
@@ -218,16 +226,16 @@ async function getDefaultChainConfigs(configs) {
     throw new Error('Can not find free port');
   }
 
-  const tendminRpcPort = await getPort({ port: getPort.makeRange(22001, 24000) });
-  const tendmintGrpcPort = await getPort({ port: getPort.makeRange(26001, 27000) });
-  const tendmintP2pPort = await getPort({ port: getPort.makeRange(27001, 28000) });
+  const tendermintRpcPort = await getPort({ port: getPort.makeRange(22001, 24000) });
+  const tendermintGrpcPort = await getPort({ port: getPort.makeRange(26001, 27000) });
+  const tendermintP2pPort = await getPort({ port: getPort.makeRange(27001, 28000) });
 
   return setConfig(configs, DEFAULT_CHAIN_NAME, {
     forgeWebPort,
     forgeGrpcPort,
-    tendminRpcPort,
-    tendmintGrpcPort,
-    tendmintP2pPort,
+    tendermintRpcPort,
+    tendermintGrpcPort,
+    tendermintP2pPort,
     workshopPort,
   });
 }
@@ -260,7 +268,11 @@ async function copyReleaseConfig(currentVersion, overwrite = true) {
  * @param {boolean} [exitOn404=true]
  * @returns
  */
-async function ensureForgeRelease(args, exitOn404 = true) {
+async function ensureForgeRelease(
+  args,
+  exitOn404 = true,
+  chainName = process.env.FORGE_CURRENT_CHAIN
+) {
   const cliConfig = {};
   const cliReleaseDir = requiredDirs.release;
 
@@ -274,15 +286,29 @@ async function ensureForgeRelease(args, exitOn404 = true) {
   const releaseYamlPath = path.join(releaseDir, './forge/release.yml');
   if (fs.existsSync(releaseDir)) {
     try {
-      const currentVersion = getForgeVersionFromYaml(releaseYamlPath);
-      const releaseYamlObj = yaml.parse(fs.readFileSync(releaseYamlPath).toString());
-      if (!releaseYamlObj || !releaseYamlObj.current) {
-        throw new Error('no current forge release selected');
+      // Read global release version
+      const curVersion = getForgeVersionFromYaml(releaseYamlPath, 'current');
+      if (!semver.valid(curVersion)) {
+        throw new Error(`no valid version field found in release config ${releaseYamlPath}`);
       }
+      cliConfig.globalVersion = curVersion;
 
-      cliConfig.currentVersion = currentVersion;
+      // Read chain-wise version
+      const chainConfigPath = getChainConfigPath(chainName);
+      const currentVersion = getForgeVersionFromYaml(chainConfigPath, 'version');
+      if (semver.valid(currentVersion)) {
+        cliConfig.currentVersion = currentVersion;
+
+        // FIXME: make this work
+        // cliConfig.configPath = getForgeVersionFromYaml(chainConfigPath, 'config');
+        debug('ensureForgeRelease.readChainConfig', chainConfigPath);
+      } else {
+        // Write chain-wise config to use global version
+        cliConfig.currentVersion = cliConfig.globalVersion;
+        updateChainConfig(chainName, { version: cliConfig.globalVersion });
+      }
     } catch (err) {
-      debug.error(err);
+      debug.error('ensureForgeRelease.readConfig.error', err);
       if (exitOn404) {
         printError(`config file ${releaseYamlPath} invalid`);
         process.exit(1);
@@ -368,14 +394,14 @@ async function ensureForgeRelease(args, exitOn404 = true) {
 }
 
 function readChainConfig(chainName) {
-  return TOML.parse(fs.readFileSync(getProfileReleaseFilePath(chainName)).toString());
+  return TOML.parse(fs.readFileSync(getChainReleaseFilePath(chainName)).toString());
 }
 
 module.exports = {
   copyReleaseConfig,
   ensureForgeRelease,
   getDefaultChainConfigs,
-  setConfigToProfile,
+  setConfigToChain,
   setFilePathOfConfig,
   readChainConfig,
 };