+++ src/core/forge-fs.js
@@ -3,6 +3,7 @@ const path = require('path');
 const os = require('os');
 const shell = require('shelljs');
 const yaml = require('yaml');
+const semver = require('semver');
 
 const debug = require('core/debug')('forge-fs');
 const {
@@ -28,11 +29,11 @@ const requiredDirs = {
 function clearDataDirectories(chainName = process.env.FORGE_CURRENT_CHAIN) {
   printWarning('Cleaning up chain data!');
 
-  const dir = getProfileDirectory(chainName);
+  const dir = getChainDirectory(chainName);
   shell.exec(`rm -rf ${dir}`);
   printInfo(`rm -f ${dir}`);
 
-  printSuccess('Data profiles cleaned!');
+  printSuccess('Chain data cleaned!');
 }
 
 function isDirectory(x) {
@@ -55,15 +56,16 @@ function getAllAppDirectories() {
 function getAllChainNames() {
   return getAllAppDirectories()
     .map(name => name.slice(name.indexOf('_') + 1))
-    .sort(chainSortHandler);
+    .sort(chainSortHandler)
+    .map(x => [x, getChainConfig(x)]);
 }
 
 function getCurrentWorkingDirectory() {
-  return process.env.CURRENT_WORKING_PROFILE;
+  return process.env.CURRENT_WORKING_CHAIN;
 }
 
 function getDataDirectory(chainName = process.env.FORGE_CURRENT_CHAIN) {
-  return path.join(getProfileDirectory(chainName), CHAIN_DATA_PATH_NAME);
+  return path.join(getChainDirectory(chainName), CHAIN_DATA_PATH_NAME);
 }
 
 function getCurrentReleaseFilePath() {
@@ -135,9 +137,19 @@ function getStorageEnginePath(version) {
 }
 
 function getForgeBinPath(version) {
-  debug('getForgeBinPath');
+  debug('getForgeBinPath, version:', version);
 
   return path.join(getForgeReleaseDirectory(), version, 'bin', 'forge');
+}
+
+function getGlobalForgeVersion() {
+  const filePath = path.join(getReleaseDirectory('forge'), 'release.yml');
+  const curVersion = getForgeVersionFromYaml(filePath, 'current');
+  if (semver.valid(curVersion)) {
+    return curVersion;
+  }
+
+  return '';
 }
 
 function isForgeBinExists(version) {
@@ -162,101 +174,156 @@ function ensureRootConfigDirectory() {
   return dir;
 }
 
-function getProfileDirectory(chainName = process.env.FORGE_CURRENT_CHAIN) {
+function getChainDirectory(chainName = process.env.FORGE_CURRENT_CHAIN) {
   const forgeRootDir = ensureRootConfigDirectory();
-  const forgeProfileDir = path.join(forgeRootDir, `forge_${chainName}`);
+  const forgeChainDir = path.join(forgeRootDir, `forge_${chainName}`);
 
-  return forgeProfileDir;
+  return forgeChainDir;
 }
 
-function getProfileReleaseFilePath(chainName = process.env.FORGE_CURRENT_CHAIN) {
-  return path.join(getProfileDirectory(chainName), 'forge_release.toml');
+function getChainReleaseFilePath(chainName = process.env.FORGE_CURRENT_CHAIN) {
+  return path.join(getChainDirectory(chainName), 'forge_release.toml');
 }
 
-function getProfileKeyFilePath(chainName = process.env.FORGE_CURRENT_CHAIN) {
-  return path.join(getProfileDirectory(chainName), 'keys');
+function getChainKeyFilePath(chainName = process.env.FORGE_CURRENT_CHAIN) {
+  return path.join(getChainDirectory(chainName), 'keys');
 }
 
-function getProfileReleaseDirectory(chainName = process.env.FORGE_CURRENT_CHAIN) {
-  return path.join(getProfileDirectory(chainName), 'forge_release');
+function getChainReleaseDirectory(chainName = process.env.FORGE_CURRENT_CHAIN) {
+  return path.join(getChainDirectory(chainName), 'forge_release');
 }
 
-function getProfileWorkshopDirectory(chainName = process.env.FORGE_CURRENT_CHAIN) {
-  return path.join(getProfileDirectory(chainName), 'workshop');
+function getChainWorkshopDirectory(chainName = process.env.FORGE_CURRENT_CHAIN) {
+  return path.join(getChainDirectory(chainName), 'workshop');
 }
 
-function createNewProfile(chainName = process.env.FORGE_CURRENT_CHAIN) {
-  const profileDirectory = getProfileDirectory(chainName);
-  if (fs.existsSync(profileDirectory)) {
-    printError(`The config file ${profileDirectory} has already exists in your current directory.`);
+function createNewChain(chainName = process.env.FORGE_CURRENT_CHAIN) {
+  const chainDirectory = getChainDirectory(chainName);
+  if (fs.existsSync(chainDirectory)) {
+    printError(`The config file ${chainDirectory} has already exists in your current directory.`);
 
     process.exit(1);
   }
 
   fs.mkdirSync(getDataDirectory(chainName), { recursive: true });
-  fs.mkdirSync(path.join(profileDirectory, 'keys'), { recursive: true });
-  print(`Initialized an empty storage space in ${profileDirectory}`);
+  fs.mkdirSync(path.join(chainDirectory, 'keys'), { recursive: true });
+  updateChainConfig(chainName, { version: getGlobalForgeVersion() });
+  print(`Initialized a new empty chain in ${chainDirectory}`);
 }
 
-function ensureProfileDirectory(chainName = process.env.FORGE_CURRENT_CHAIN) {
-  const forgeProfileDir = getProfileDirectory(chainName);
+function ensureChainDirectory(chainName = process.env.FORGE_CURRENT_CHAIN) {
+  const forgeChainDir = getChainDirectory(chainName);
 
-  if (!fs.existsSync(forgeProfileDir)) {
-    fs.mkdirSync(forgeProfileDir, { recursive: true });
+  if (!fs.existsSync(forgeChainDir)) {
+    fs.mkdirSync(forgeChainDir, { recursive: true });
     fs.mkdirSync(getDataDirectory(chainName), { recursive: true });
-    fs.mkdirSync(path.join(forgeProfileDir, 'keys'), { recursive: true });
-    print(`Initialized an empty storage space in ${forgeProfileDir}`);
+    fs.mkdirSync(path.join(forgeChainDir, 'keys'), { recursive: true });
+    updateChainConfig(chainName, { version: getGlobalForgeVersion() });
+    print(`Initialized an empty storage space in ${forgeChainDir}`);
   }
 
-  return forgeProfileDir;
+  return forgeChainDir;
 }
 
 function getLogfile(chainName = process.env.FORGE_CURRENT_CHAIN, fileName) {
-  return path.join(getProfileReleaseDirectory(chainName), 'core', 'logs', fileName);
+  return path.join(getChainReleaseDirectory(chainName), 'core', 'logs', fileName);
 }
 
-function getForgeVersionFromYaml(yamlPath) {
+function getForgeVersionFromYaml(yamlPath, key = 'current') {
   try {
     if (!fs.existsSync(yamlPath)) {
       return '';
     }
 
     const releaseYamlObj = yaml.parse(fs.readFileSync(yamlPath).toString());
-    if (!releaseYamlObj || !releaseYamlObj.current) {
+    if (!releaseYamlObj || !releaseYamlObj[key]) {
       throw new Error('no current forge release selected');
     }
 
-    return releaseYamlObj.current;
+    return releaseYamlObj[key];
   } catch (err) {
     debug.error(err);
     return '';
   }
 }
 
-function getCurrentForgeVersion() {
-  const filePath = path.join(getForgeReleaseDirectory(), 'release.yml');
+function updateReleaseYaml(asset, version) {
+  try {
+    const filePath = path.join(getReleaseDirectory(asset), 'release.yml');
+    shell.exec(`touch ${filePath}`, { silent: true });
+    debug('updateReleaseYaml', { asset, version, filePath });
+    const yamlObj = fs.existsSync(filePath)
+      ? yaml.parse(fs.readFileSync(filePath).toString()) || {}
+      : {};
+    if (yamlObj.current) {
+      yamlObj.old = yamlObj.current;
+    }
+    yamlObj.current = version;
+    fs.writeFileSync(filePath, yaml.stringify(yamlObj), { flag: 'w+' });
+  } catch (err) {
+    printError(err);
+  }
+}
 
-  return getForgeVersionFromYaml(filePath);
+function getChainConfigPath(chainName) {
+  return path.join(getChainDirectory(chainName), 'config.yml');
+}
+
+function getChainConfig(chainName) {
+  const filePath = path.join(getChainDirectory(chainName), 'config.yml');
+  try {
+    return fs.existsSync(filePath) ? yaml.parse(fs.readFileSync(filePath).toString()) || {} : {};
+  } catch (err) {
+    // Do nothing
+    return {};
+  }
+}
+
+/**
+ * Read config from yaml file.
+ * @param {string} filePath
+ * @returns {json} return config of json format, if config is empty, return empty json object.
+ */
+function readYamlConfig(filePath) {
+  const yamlObj = fs.existsSync(filePath)
+    ? yaml.parse(fs.readFileSync(filePath).toString()) || {}
+    : {};
+
+  return yamlObj;
+}
+
+function updateChainConfig(chainName, config = {}) {
+  try {
+    const filePath = getChainConfigPath(chainName);
+    debug('updateChainConfig', { chainName, config });
+
+    const chainConfig = readYamlConfig(filePath);
+    fs.writeFileSync(filePath, yaml.stringify(Object.assign(chainConfig, config)), { flag: 'w+' });
+  } catch (err) {
+    printError(err);
+  }
 }
 
 module.exports = {
   clearDataDirectories,
-  createNewProfile,
-  ensureProfileDirectory,
+  createNewChain,
+  ensureChainDirectory,
   getAllAppDirectories,
   getAllChainNames,
   getCliDirectory,
   getConsensusEnginBinPath,
   getCurrentReleaseFilePath,
-  getCurrentForgeVersion,
+  getDataDirectory,
   getForgeBinPath,
-  getProfileDirectory,
+  getChainDirectory,
+  getChainConfigPath,
+  getChainConfig,
   getForgeSimulatorReleaseDirectory,
   getForgeWebReleaseDirectory,
   getForgeWorkshopReleaseDirectory,
-  getProfileReleaseFilePath,
-  getDataDirectory,
-  getProfileWorkshopDirectory,
+  getGlobalForgeVersion,
+  getChainReleaseFilePath,
+  getChainWorkshopDirectory,
   getLogfile,
   getRootConfigDirectory,
   getTendermintHomeDir,
@@ -264,9 +331,11 @@ module.exports = {
   getOriginForgeReleaseFilePath,
   getForgeReleaseDirectory,
   getStorageEnginePath,
-  getProfileKeyFilePath,
+  getChainKeyFilePath,
   isForgeBinExists,
   isDirectory,
   isFile,
   requiredDirs,
+  updateReleaseYaml,
+  updateChainConfig,
 };